fix create update route blog

+++ app/utils/functions.js
@@ -69,8 +69,10 @@ function VerifyRefreshToken(token){
 }
 
 function deleteFileInPublic(fileAddress){
-const pathFile=path.join(__dirname,'..','..','public',fileAddress);
- fs.unlinkSync(pathFile)   
+    if(fileAddress){
+        const pathFile=path.join(__dirname,'..','..','public',fileAddress);
+if(fs.existsSync(pathFile)) fs.unlinkSync(pathFile) 
+    }  
 }
 module.exports={
     RandomNumberGenerator,
